divide into reusable components
